I added four new pages, the registration page, login page, the page for password reset, and the page for creating a new password. I have the four pages set up as components, and in the routes folder, for the four pages, I call the components within the return statement for the page they should be linked to. I added a logo in the top left, and the input and button components are from shadcn.
Created a hidden path folder that contains all the routes that need LoginNav. I put the LoginNav within the route.tsx of that folder. I also moved the page information to the root components folder, and kept the button and input within the ui folder

Added ZOD schema validation, and added a link to the register page from the login page, and vice versa.

Update route.tsx

Removed unnecessary fragment and TanStackRouterDevtools.

adding form field as a component